Added remove controller and service
