+++ controllers/index.js
@@ -5,6 +5,7 @@ const searchControllerFactory = require('./search');
 const findOneControllerFactory = require('./findone');
 const insertControllerFactory = require('./insert');
 const insertOrCountControllerFactory = require('./insertorcount');
+const removeControllerFactory = require('./remove');
 
 const controllers = (Application) => {
 	debug('Called');
@@ -97,36 +98,11 @@ const controllers = (Application) => {
 				return next(err);
 			}
 		},
-		'universal.remove': async (req, res, next) => {
-			const { _id } = req.swagger.params;
-			debug('.remove called: ', _id);
-			if (!db) {
-				throw new Error('Cant access to universal.* without MongoDB Connection');
-			}
-
-			try {
-				if (Application.hooks['*'] && Application.hooks['*'].beforeRemove) {
-					await Application.hooks['*'].beforeRemove(req, _id, Application);
-				}
-				if (Application.hooks[req.swagger.apiPath] && Application.hooks[req.swagger.apiPath].beforeRemove) {
-					await Application.hooks[req.swagger.apiPath].beforeRemove(req, _id, Application);
-				}
-
-				let removedDocument = await services.findOne(req.swagger.apiPath, { _id: db.ObjectId(_id) }, {});
-				const result = await services.remove(req.swagger.apiPath, _id);
-
-				if (Application.hooks['*'] && Application.hooks['*'].afterRemove) {
-					removedDocument = await Application.hooks['*'].afterRemove(req, { ...removedDocument, result }, Application);
-				}
-				if (Application.hooks[req.swagger.apiPath] && Application.hooks[req.swagger.apiPath].afterRemove) {
-					removedDocument = await Application.hooks[req.swagger.apiPath].afterRemove(req, { ...removedDocument, result }, Application);
-				}
-
-				return res.json({ ...removedDocument, result });
-			} catch (err) {
-				return next(err);
-			}
-		},
+		'universal.remove': removeControllerFactory({
+			db,
+			services,
+			Application,
+		}),
 		'universal.today': async (req, res, next) => {
 			debug('.today called');
 			if (!db) {
